+++ MotionTracking/motion.py
@@ -1,0 +1,60 @@
+import cv2 as cv
+
+import Common.color as Color
+
+
+def detect_motion_sparse(img, frame_1, frame_2, kernel=None, filter_blur=(5, 5), iter_dilate=4, VALUE_AREA=150):
+    r"""
+
+    """
+    frame_1 = cv.cvtColor(frame_1, cv.COLOR_GRAY2BGR)
+    frame_2 = cv.cvtColor(frame_2, cv.COLOR_GRAY2BGR)
+
+    diff = cv.absdiff(frame_1, frame_2)
+    diff_gray = cv.cvtColor(diff, cv.COLOR_BGR2GRAY)
+    blur = cv.GaussianBlur(diff_gray, filter_blur, 0)
+    _, thresh = cv.threshold(blur, 20, 255, cv.THRESH_BINARY)
+
+    dilate = cv.dilate(thresh, kernel, iterations=iter_dilate)
+
+    cv.imshow("Dilate", dilate)
+    cv.waitKey(1)
+
+    contours, _ = cv.findContours(dilate, cv.RETR_TREE, cv.CHAIN_APPROX_SIMPLE)
+
+    for cnt in contours:
+        (x, y, w, h) = cv.boundingRect(cnt)
+        area = cv.contourArea(cnt)
+
+        if area <= VALUE_AREA:
+            continue
+
+        cv.rectangle(img, (x, y), (x + w, y + h), Color.RED, 2)
+
+
+def detect_car(img, original):
+    r"""
+
+    """
+    img_gray = cv.cvtColor(img, cv.COLOR_BGR2GRAY)
+    _, thresh = cv.threshold(img_gray, 20, 255, cv.THRESH_BINARY)
+
+    kernel = cv.getStructuringElement(cv.MORPH_RECT, (5, 5))
+    opening = cv.morphologyEx(thresh, cv.MORPH_OPEN, kernel, iterations=4)
+
+    contours, _ = cv.findContours(opening, cv.RETR_EXTERNAL, cv.CHAIN_APPROX_NONE)
+
+    for num, cnt in enumerate(contours):
+        (x, y, w, h) = cv.boundingRect(cnt)
+        area = cv.contourArea(cnt)
+
+        if area <= 500:
+            continue
+
+        cv.rectangle(original, (x, y), (x + w, y + h), (0, 0, 255), 2)
+
+        height, width, _ = original.shape
+        thick = int((height + width) // 900)
+        cv.putText(original, f"Vehicle {num}", (x, y - 12), 0, 1e-3 * height, (0, 0, 255), thick)
+
+    cv.imshow("Mask motion", opening)
