--- tracking/motion.py
@@ -1,32 +1,0 @@
-import cv2 as cv
-
-import common.color as Color
-
-
-def detect(img, frame_1, frame_2, kernel=None, filter_blur=(5, 5), iter_dilate=4, VALUE_AREA=150):
-    r"""
-
-    """
-    frame_1 = cv.cvtColor(frame_1, cv.COLOR_GRAY2BGR)
-    frame_2 = cv.cvtColor(frame_2, cv.COLOR_GRAY2BGR)
-
-    diff = cv.absdiff(frame_1, frame_2)
-    diff_gray = cv.cvtColor(diff, cv.COLOR_BGR2GRAY)
-    blur = cv.GaussianBlur(diff_gray, filter_blur, 0)
-    _, thresh = cv.threshold(blur, 20, 255, cv.THRESH_BINARY)
-
-    dilate = cv.dilate(thresh, kernel, iterations=iter_dilate)
-
-    cv.imshow("Dilate", dilate)
-    cv.waitKey(1)
-
-    contours, _ = cv.findContours(dilate, cv.RETR_TREE, cv.CHAIN_APPROX_SIMPLE)
-
-    for cnt in contours:
-        (x, y, w, h) = cv.boundingRect(cnt)
-        area = cv.contourArea(cnt)
-
-        if area <= VALUE_AREA:
-            continue
-
-        cv.rectangle(img, (x, y), (x + w, y + h), Color.RED, 2)
